Added in the data export from the DB.
`data_export` in the `src` folder assumes a local SQL version of the database. This code will get modified in the future (resumably) as new datasets get added to SISAL, where we can check our Neotoma siteids against the SISAL IDs to ensure we're only uploading new records. We may also need to check for updates, but we can sort that out.